renamed files, updated tests
