+++ py/picca/tests/delta_extraction/data/absorber_mask_print.txt
