+++ py/picca/tests/delta_extraction/mask_tests.py
@@ -9,9 +9,9 @@ from picca.delta_extraction.mask import Mask
 from picca.delta_extraction.masks.lines_mask import LinesMask
 from picca.delta_extraction.masks.dla_mask import DlaMask
 from picca.delta_extraction.masks.dla_mask import defaults as defaults_dla_mask
-from picca.delta_extraction.masks.sdss_absorber_mask import SdssAbsorberMask
-from picca.delta_extraction.masks.sdss_absorber_mask import (
-    defaults as defaults_sdss_absorber_mask)
+from picca.delta_extraction.masks.absorber_mask import AbsorberMask
+from picca.delta_extraction.masks.absorber_mask import (
+    defaults as defaults_absorber_mask)
 from picca.delta_extraction.errors import MaskError
 from picca.delta_extraction.utils import setup_logger
 from picca.tests.delta_extraction.abstract_test import AbstractTest
@@ -45,14 +45,14 @@ class MaskTest(AbstractTest):
 
     def test_absorber_mask(self):
         """Test correct initialisation and inheritance for class
-        SdssAbsorberMask
+        AbsorberMask
 
-        Load a SdssAbsorberMask instace and check that it is
+        Load a AbsorberMask instace and check that it is
         correctly initialized.
         """
         in_file = f"{THIS_DIR}/data/dummy_absorbers_cat.fits.gz"
-        out_file = f"{THIS_DIR}/results/sdss_absorber_mask_print.txt"
-        test_file = f"{THIS_DIR}/data/sdss_absorber_mask_print.txt"
+        out_file = f"{THIS_DIR}/results/absorber_mask_print.txt"
+        test_file = f"{THIS_DIR}/data/absorber_mask_print.txt"
 
         # setup printing
         setup_logger(log_file=out_file)
@@ -60,10 +60,10 @@ class MaskTest(AbstractTest):
         # initialize mask
         config = ConfigParser()
         config.read_dict({"mask": {"filename": in_file}})
-        for key, value in defaults_sdss_absorber_mask.items():
+        for key, value in defaults_absorber_mask.items():
             if key not in config["mask"]:
                 config["mask"][key] = str(value)
-        mask = SdssAbsorberMask(config["mask"])
+        mask = AbsorberMask(config["mask"])
         self.assertTrue(isinstance(mask, Mask))
         self.assertTrue(mask.absorber_mask_width == 2.5)
 
@@ -103,10 +103,10 @@ class MaskTest(AbstractTest):
         config = ConfigParser()
         config.read_dict({"mask": {"filename": in_file,
                                    "absorber mask width": 1.5,}})
-        for key, value in defaults_sdss_absorber_mask.items():
+        for key, value in defaults_absorber_mask.items():
             if key not in config["mask"]:
                 config["mask"][key] = str(value)
-        mask = SdssAbsorberMask(config["mask"])
+        mask = AbsorberMask(config["mask"])
         self.assertTrue(mask.absorber_mask_width == 1.5)
 
         reset_logger()
